hack: more random freebsd things
i still can't get it to even start gathering data from the provider
i'm testing (linux.process_stats), but more of this is compiling.

+++ include/perfetto/ext/base/metatrace.h
@@ -134,8 +134,8 @@ struct Record {
   uint32_t timestamp_ns_low = 0;
 
   // [adrian] shouldn't this be 64 bit?
-  uint64_t thread_id = 0;
-  //uint32_t thread_id = 0;
+  //uint64_t thread_id = 0;
+  uint32_t thread_id = 0;
 
   union {
     // Only one of the two elements can be zero initialized, clang complains
